Code Suggestions
Keep in mind many of these are very subjective but hopefully you find
them helpful!

I have rearranged the code slightly to make it more in line with the
names used for the modules. `pricing_rules` now actually contains the
functions used to apply the BOGOF and bulk discount rules. I explicitly
defined the rules dict in `test_checkout` since I imagine this is how an
end-user might apply the rules.

I'm a big fan of using dicts in place of `if`/`elif` chains to make
adding new "branches" easier although some might say the latter is more
pythonic.

Apart from that I don't really have any suggestions. It was already
excellent Python code. OOP should be thought of a bit differently in
Python when compared with Ruby or Java. It's perfectly OK to group
things in modules without them having to be inside a class.